pyproject.toml
[project]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
requires-python = ">=3.11"
dependencies = [
    "bing-image-downloader>=1.1.2",
    "boto3>=1.37.25",
    "flask>=3.1.0",
    "google-images-download>=2.8.0",
    "numpy>=2.2.4",
    "pillow>=11.2.0",
    "pymongo>=4.11.3",
    "requests>=2.32.3",
    "voyageai>=0.3.2",
]
